+++ tests/mcp/rag-handler.test.js
@@ -4,7 +4,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-vi.mock("../../src/rag/embedder.js", () => {
+vi.mock("karajan-core/rag/embedder", () => {
   class FakeEmbedder {
     constructor() { this.dim = 8; }
     async embed() { const v = new Float32Array(8); v[0] = 1; return v; }
