+++ tests/orchestrator/rag-context-stage.test.js
@@ -4,7 +4,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-vi.mock("../../src/rag/embedder.js", () => {
+vi.mock("karajan-core/rag/embedder", () => {
   class FakeEmbedder {
     constructor() { this.dim = 768; }
     async embed() { const v = new Float32Array(768); v[0] = 1; return v; }
