+++ tests/rag/commands.test.js
@@ -6,7 +6,9 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 const noopLogger = { info: vi.fn(), warn: vi.fn(), error: vi.fn() };
 
-vi.mock("../../src/rag/embedder.js", () => {
+// KJC-TSK-0632: the embedder moved to karajan-core/rag — mock the REAL
+// module (the src/rag path is now a re-export shim the consumers bypass).
+vi.mock("karajan-core/rag/embedder", () => {
   class FakeEmbedder {
     constructor() { this.dim = 8; }
     async embed() { const v = new Float32Array(8); v[0] = 1; return v; }
